+++ src/runtime/mod.rs
@@ -5,8 +5,8 @@
 
 pub mod force;
 mod gss;
-pub mod gss_acc;
-pub mod leveled_gss;
+pub(crate) mod gss_acc;
+pub(crate) mod leveled_gss;
 mod mask;
 mod state;